# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Moroni's Quest Application",
    "description": "API documentation for Moronis Quest event management system. Provides endpoints for youth and leader registration, Book of Mormon challenges, photo gallery, and inquiry management for the Sherwood Park, Alberta stake.",
    "version": "1.0.0"
  },
  "host": "moroni-quest.onrender.com",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/auth/user": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
...
}